requirements.txt
# Após instalar as dependências, rode: python -m spacy download de_core_news_lg
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
spacy>=3.7,<4.0
openai>=1.40,<2.0
python-dotenv>=1.0,<2.0
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.2,<3.0
whitenoise>=6.7,<7.0
gunicorn>=22.0,<24.0